# docs/openapi/relay.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AI模型接口",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/kling/v1/videos/text2video": {
      "post": {
        "summary": "Kling 文生视频",
        "deprecated": false,
        "description": "使用 Kling 模型从文本描述生成视频。\n\n支持的模型：kling-v1, kling-v1-5 等\n",
        "operationId": "createKlingText2Video",
        "tags": [
          "视频生成/Kling格式"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员站起身走了",
                "duration": 5,
                "width": 1280,
                "height": 720,
                "fps": 30
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kling/v1/videos/text2video/{task_id}": {
      "get": {
        "summary": "获取 Kling 文生视频任务状态",
        "deprecated": false,
        "description": "查询 Kling 文生视频任务的状态和结果。",
        "operationId": "getKlingText2Video",
        "tags": [
          "视频生成/Kling格式"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "任务 ID",
            "required": true,
            "example": "task-abc123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kling/v1/videos/image2video": {
      "post": {
        "summary": "Kling 图生视频",
        "deprecated": false,
        "description": "使用 Kling 模型从图片生成视频。\n\n支持通过 image 参数传入图片 URL 或 Base64 编码的图片数据。\n",
        "operationId": "createKlingImage2Video",
        "tags": [
          "视频生成/Kling格式"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "人物转身走开",
                "image": "https://example.com/image.jpg",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/kling/v1/videos/image2video/{task_id}": {
      "get": {
        "summary": "获取 Kling 图生视频任务状态",
        "deprecated": false,
        "description": "查询 Kling 图生视频任务的状态和结果。",
        "operationId": "getKlingImage2Video",
        "tags": [
          "视频生成/Kling格式"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "任务 ID",
            "required": true,
            "example": "task-abc123",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/video/generations": {
      "post": {
        "summary": "创建视频生成任务",
        "deprecated": false,
        "description": "提交视频生成任务，支持文生视频和图生视频。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "kling-v1",
                "prompt": "宇航员在月球上漫步",
                "duration": 5,
                "width": 1280,
                "height": 720
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/video/async-generations": {
      "post": {
        "summary": "创建视频异步生成任务",
        "deprecated": false,
        "description": "提交视频异步生成任务，兼容 sora2 下游调用格式。\n\n返回任务 ID，可通过 GET 接口查询任务状态。\n",
        "operationId": "createVideoAsyncGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              },
              "example": {
                "model": "sora2",
                "prompt": "一个美食节目镜头，厨师把意大利面装盘，暖色灯光，镜头平稳推进，无文字，无logo",
                "duration": 4,
                "aspect_ratio": "16:9",
                "async": true
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功创建视频异步生成任务",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/video/generations/{task_id}": {
      "get": {
        "summary": "获取视频生成任务状态",
        "deprecated": false,
        "description": "查询视频生成任务的状态和结果。\n\n任务状态：\n- `queued`: 排队中\n- `in_progress`: 生成中\n- `completed`: 已完成\n- `failed`: 失败\n",
        "operationId": "getVideoGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "任务 ID",
            "required": true,
            "example": "abcd1234efgh",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/video/async-generations/{task_id}": {
      "get": {
        "summary": "获取视频异步生成任务状态",
        "deprecated": false,
        "description": "查询视频异步生成任务的状态和结果。\n\n任务状态：\n- `queued`: 排队中\n- `in_progress`: 生成中\n- `completed`: 已完成\n- `failed`: 失败\n",
        "operationId": "getVideoAsyncGeneration",
        "tags": [
          "视频生成"
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "description": "任务 ID",
            "required": true,
            "example": "abcd1234efgh",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取任务状态",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTaskResponse"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "description": "错误信息"
              },
              "type": {
                "type": "string",
                "description": "错误类型"
              },
              "param": {
                "type": "string",
                "description": "相关参数",
                "nullable": true
              },
              "code": {
                "type": "string",
                "description": "错误代码",
                "nullable": true
              }
            }
          }
        }
      },
...
      "VideoRequest": {
        "type": "object",
        "description": "视频生成请求",
        "properties": {
          "model": {
            "type": "string",
            "description": "模型/风格 ID",
            "example": "kling-v1"
          },
          "prompt": {
            "type": "string",
            "description": "文本描述提示词",
            "example": "宇航员站起身走了"
          },
          "image": {
            "type": "string",
            "description": "图片输入 (URL 或 Base64)",
            "example": "https://example.com/image.jpg"
          },
          "duration": {
            "type": "number",
            "description": "视频时长（秒）",
            "example": 5
          },
          "width": {
            "type": "integer",
            "description": "视频宽度",
            "example": 1280
          },
          "height": {
            "type": "integer",
            "description": "视频高度",
            "example": 720
          },
          "fps": {
            "type": "integer",
            "description": "视频帧率",
            "example": 30
          },
          "seed": {
            "type": "integer",
            "description": "随机种子",
            "example": 20231234
          },
          "n": {
            "type": "integer",
            "description": "生成视频数量",
            "example": 1
          },
          "response_format": {
            "type": "string",
            "description": "响应格式",
            "example": "url"
          },
          "user": {
            "type": "string",
            "description": "用户标识",
            "example": "user-1234"
          },
          "metadata": {
            "type": "object",
            "description": "扩展参数 (如 negative_prompt, style, quality_level 等)",
            "additionalProperties": true,
            "properties": {}
          }
        }
      },
      "VideoResponse": {
        "type": "object",
        "description": "视频生成任务提交响应",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任务 ID",
            "example": "abcd1234efgh"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "example": "queued"
          }
        }
      },
      "VideoTaskResponse": {
        "type": "object",
        "description": "视频任务状态查询响应",
        "properties": {
          "task_id": {
            "type": "string",
            "description": "任务 ID",
            "example": "abcd1234efgh"
          },
          "status": {
            "type": "string",
            "description": "任务状态",
            "enum": [
              "queued",
              "in_progress",
              "completed",
              "failed"
            ],
            "example": "completed"
          },
          "url": {
            "type": "string",
            "description": "视频资源 URL（成功时）",
            "example": "https://example.com/video.mp4"
          },
          "format": {
            "type": "string",
            "description": "视频格式",
            "example": "mp4"
          },
          "metadata": {
            "$ref": "#/components/schemas/VideoTaskMetadata"
          },
          "error": {
            "$ref": "#/components/schemas/VideoTaskError"
          }
        }
      },
      "VideoTaskMetadata": {
        "type": "object",
        "description": "视频任务元数据",
        "properties": {
          "duration": {
            "type": "number",
            "description": "实际生成的视频时长",
            "example": 5
          },
          "fps": {
            "type": "integer",
            "description": "实际帧率",
            "example": 30
          },
          "width": {
            "type": "integer",
            "description": "实际宽度",
            "example": 1280
          },
          "height": {
            "type": "integer",
            "description": "实际高度",
            "example": 720
          },
          "seed": {
            "type": "integer",
            "description": "使用的随机种子",
            "example": 20231234
          }
        }
      },
      "VideoTaskError": {
        "type": "object",
        "description": "视频任务错误信息",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误信息"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "使用 Bearer Token 认证。\n格式: `Authorization: Bearer sk-xxxxxx`\n"
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session认证，通过登录接口获取"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Access Token认证，格式: Bearer {access_token}，通过 /api/user/token 接口生成"
      },
      "NewApiUser": {
        "type": "apiKey",
        "in": "header",
        "name": "New-Api-User",
        "description": "用户ID请求头，必须与当前登录用户ID匹配，使用Session或AccessToken认证时必须提供"
      },
      "Combination": {
        "group": [
          {
            "id": "SessionAuth"
          },
          {
            "id": "NewApiUser"
          }
        ],
        "type": "combination"
      },
      "Combination1": {
        "group": [
          {
            "id": "AccessToken"
          },
          {
...
  },
  "servers": [],
  "security": [
    {
      "BearerAuth": []
    }
  ]
}